Struct influxdb::Client

pub struct Client { /* private fields */ }

Internal Representation of a Client

Implementations

impl Client

pub fn new<S1, S2>(url: S1, database: S2) -> Selfwhere
S1: Into<String>,
S2: Into<String>,

Instantiates a new Client

Arguments

url: The URL where InfluxDB is running (ex. http://localhost:8086).

database: The Database against which queries and writes will be run.

Examples

use influxdb::Client;

let _client = Client::new("http://localhost:8086", "test");

pub fn with_auth<S1, S2>(self, username: S1, password: S2) -> Selfwhere
S1: Into<String>,
S2: Into<String>,

Add authentication/authorization information to Client

Arguments

username: The Username for InfluxDB.

password: The Password for the user.

Examples

use influxdb::Client;

let _client = Client::new("http://localhost:9086", "test").with_auth("admin", "password");

pub fn with_http_client(self, http_client: HttpClient) -> Self

Replaces the HTTP Client

pub fn with_token<S>(self, token: S) -> Selfwhere
S: Into<String>,

Add authorization token to Client

This is designed for influxdb 2.0’s backward-compatible API which requires authrozation by default. You can create such token from console of influxdb 2.0 .

pub fn database_name(&self) -> &str

Returns the name of the database the client is using

pub fn database_url(&self) -> &str

Returns the URL of the InfluxDB installation the client is using

pub async fn ping(&self) -> Result<(String, String), Error>

Pings the InfluxDB Server

Returns a tuple of build type and version number

pub async fn query<Q>(&self, q: Q) -> Result<String, Error>where
Q: Query,

Sends a ReadQuery or WriteQuery to the InfluxDB Server.

A version capable of parsing the returned string is available under the serde_integration

Arguments

q: Query of type ReadQuery or WriteQuery

Examples

use influxdb::{Client, Query, Timestamp};
use influxdb::InfluxDbWriteable;
use std::time::{SystemTime, UNIX_EPOCH};

let start = SystemTime::now();
let since_the_epoch = start
  .duration_since(UNIX_EPOCH)
  .expect("Time went backwards")
  .as_millis();

let client = Client::new("http://localhost:8086", "test");
let query = Timestamp::Milliseconds(since_the_epoch)
    .into_query("weather")
    .add_field("temperature", 82);
let results = client.query(query).await?;

Errors

If the function can not finish the query, a Error variant will be returned.

impl Client

pub async fn json_query(
&self,
q: ReadQuery
) -> Result<DatabaseQueryResult, Error>